Bug 1392606 - Fix MIPS64 simulator build compilation failures. r=bbouvier
UltraBlame original commit: 34a8455c373ccd53d11dd881c5177a6937cdadf0

+++ js/src/wasm/WasmStubs.cpp
@@ -1174,7 +1174,7 @@ wasm::GenerateInterruptExit(MacroAssembler& masm, Label* throwLabel)
     
     masm.addToStackPtr(Imm32(2 * sizeof(intptr_t)));
     masm.as_jr(HeapReg);
-    masm.loadPtr(Address(StackPointer, -sizeof(intptr_t)), HeapReg);
+    masm.loadPtr(Address(StackPointer, -int32_t(sizeof(intptr_t))), HeapReg);
 #elif defined(JS_CODEGEN_ARM)
     {
